backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas==2.2.2
networkx==3.3
groq==0.9.0
python-dotenv==1.0.1
openpyxl==3.1.2
aiofiles==23.2.1
pyarrow==16.1.0
